+++ core/.projen/tasks.json
@@ -41,6 +41,12 @@
         },
         {
           "spawn": "docgen"
+        },
+        {
+          "exec": "npx projen copy-resources"
+        },
+        {
+          "exec": "npx projen pip-install"
         }
       ]
     },
@@ -50,6 +56,12 @@
       "steps": [
         {
           "exec": "tsc --noEmit --project tsconfig.jest.json"
+        },
+        {
+          "exec": "npx projen copy-resources"
+        },
+        {
+          "exec": "npx projen pip-install"
         }
       ]
     },
@@ -271,6 +283,36 @@
       "steps": [
         {
           "exec": "cdk destroy --app=./lib/integ.default.js"
+        }
+      ]
+    },
+    "copy-resources": {
+      "name": "copy-resources",
+      "description": "Copy all resources directories from src to lib",
+      "steps": [
+        {
+          "exec": "cp -r src/data-generator/resources lib/data-generator"
+        },
+        {
+          "exec": "cp -r src/synchronous-athena-query/resources lib/synchronous-athena-query"
+        },
+        {
+          "exec": "cp -r src/synchronous-crawler/resources lib/synchronous-crawler"
+        }
+      ]
+    },
+    "pip-install": {
+      "name": "pip-install",
+      "description": "pip install all folders in lib that has requirements.txt",
+      "steps": [
+        {
+          "exec": "pip install -r lib/data-generator/resources/lambdas/setup/requirements.txt --target lib/data-generator/resources/lambdas/setup"
+        },
+        {
+          "exec": "pip install -r lib/synchronous-athena-query/resources/lambdas/requirements.txt --target lib/synchronous-athena-query/resources/lambdas"
+        },
+        {
+          "exec": "pip install -r lib/synchronous-crawler/resources/lambdas/requirements.txt --target lib/synchronous-crawler/resources/lambdas"
         }
       ]
     },
