+++ core/package.json
@@ -27,6 +27,8 @@
     "docgen": "npx projen docgen",
     "test:deploy": "npx projen test:deploy",
     "test:destroy": "npx projen test:destroy",
+    "copy-resources": "npx projen copy-resources",
+    "pip-install": "npx projen pip-install",
     "release": "npx projen release",
     "projen": "npx projen"
   },
@@ -46,6 +48,7 @@
     "eslint-import-resolver-node": "^0.3.6",
     "eslint-import-resolver-typescript": "^2.5.0",
     "eslint-plugin-import": "^2.24.2",
+    "glob": "^7.2.0",
     "jest": "^27.2.2",
     "jest-junit": "^12",
     "jsii": "^1.34.0",
