+++ core/src/data-generator/index.ts
@@ -1,0 +1,1 @@
+export { DataGeneratorProps, DataGenerator } from './data-generator';
