+++ core/src/data-generator/resources/lambdas/setup/lambda.py
