+++ core/src/synchronous-athena-query/resources/lambdas/requirements.txt
