+++ core/src/synchronous-athena-query/synchronous-athena-query.ts
